autodetect dotenv and load it if available
